Added FCM API for sending messages
Okay, so here is the meat of the FCM API implementation minus one key
thing: error handling for server-side errors. This CL is already big
and the error handling is complex enough that I want to reserve it for
its own CL. So, there are a few TODOs in this CL which will be resolved
in the next (and last) CL.

Also, I know test/messaging.spec.ts is quite large, but there is a lot
to test and I’m pretty happy with the 99.38% test coverage for the FCM
methods. I tried to organize the tests in as readable a manner as
possible, but please let me know if you have suggestions.

Change-Id: I94a6a0b8bc952ae828c501135f0b6ea1f7c3c0b4